[V3] Fixing lmdb benchmark compile error and json-comments-rs warnings (#44)

+++ crates/lmdb-js-lite/benches/lmdb_js_safe_benchmark.rs
@@ -11,7 +11,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     DatabaseWriter::new(&LMDBOptions {
       path: "benchmark-databases/test.db".to_string(),
       async_writes: false,
-      map_size: Some(1024 * 1024 * 1024),
+      map_size: Some((1024 * 1024 * 1024) as f64),
     })
     .unwrap()
   };
